FundRefProj/requirements.txt
asgiref==3.4.1
blis==0.2.4
certifi==2021.10.8
chardet==4.0.0
click==7.1.2
cymem==2.0.6
Django==3.2
djangorestframework==3.12.4
Flask==1.1.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==2.0.1
murmurhash==1.0.6
numpy==1.19.5
pandas==1.1.5
plac==0.9.6
preshed==2.0.1
python-dateutil==2.8.2
pytz==2021.3
requests==2.25.1
six==1.16.0
spacy==2.1.9
sqlparse==0.4.2
srsly==1.0.5
thinc==7.0.8
tqdm==4.59.0
typing-extensions==3.10.0.2
urllib3==1.26.7
wasabi==0.8.2
Werkzeug==1.0.1
